Allow DOCKER_HOST to use the SSH scheme (#190)
* Allow DOCKER_HOST to use the SSH scheme

This commit pulls in the Docker CLI's connhelper package which provides some
wrapper code for creating a client that can communicate with a Docker instance
over SSH.

The connection helper package is being used by adding a new ClientOpts()
function that will return a slice of client.Opt values appropriate for
connecting to DOCKER_HOST. If DOCKER_HOST is an ssh url, a set of opts built
using connhelper will be returned. If not, a set containing only client.FromEnv
will be returned, preserving the existing behavior of ctlptl.

* Include client.FromEnv at the beginning of SSH client options.

This ensures that the Docker version is still read. It also makes the code a
little easier to read.