cluster: don't use registry mirroring on minikube
Fixes https://github.com/tilt-dev/ctlptl/issues/180

This changes the local-registry-hosting config so that,
instead of pushing and pulling from the same URL (localhost:5000),
we'll push to localhost:5000 and pull from [registry-name]:5000.

To maximize compatibility, we'll still add the registry mirroring
to the containerd.toml. But because minikube doesn't do a good job
preserving the toml across stop/start, the local-registry-hosting
config will tell tilt not to use this feature.