registry: add support for custom image
The most common use case here is for restricted environments that
don't have access to Docker hub. (That said, IMO it's better to
use the built-in `--registry-mirrors` option in Docker in this
situation.)

Unfortunately, listing registries is currently done by looking for
the ancestor image of `registry:2`, which won't work when using a
custom image. As a result, a new label is applied to all registries:
`dev.tilt.ctlptl.role=registry`. The label key uses the reverse DNS
convention as suggested by Docker[1]. For backwards compatibility,
if _no_ registries with the label are found, we'll do a search by
ancestor.

NOTE: Because of this, the "desired" registry labels won't match
and there will be a one-time re-creation of the registry the first
time `ctlptl apply` is run against a registry created with an older
version of ctlptl.

Fixes #215.