registry: add support for custom image (#217)
The most common use case here is for restricted environments that
don't have access to Docker hub. (That said, IMO it's better to
use the built-in `--registry-mirrors` option in Docker in this
situation.)

Currently, listing registries is currently done by looking for the
the ancestor image of `registry:2`, which won't work when using a
custom image. As a result, a new label is applied to all registries
created by `ctltptl`: `dev.tilt.ctlptl.role=registry`. The label key
uses the reverse DNS convention as suggested by Docker[1].

For both backwards compatibility (and interop with other tools that
create clusters such as `k3d`), we'll still search for registries by
`ancestor` as well. Additionally, the label will _not_ be applied to
any existing `ctlptl`-managed registries unless they're re-created
for another reason, as labels on containers are immutable, and there's
no compelling reason to forcibly re-create users' registries to add
the label.

Fixes #215.